requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.10.5
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blis==0.7.11
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cryptography==44.0.1
cymem==2.0.11
dnspython==2.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
fastapi==0.115.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
h11==0.14.0
huggingface-hub==0.29.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kafka-python==2.0.2
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.1.0
preshed==3.0.9
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.8.0
pypdfium2==4.30.1
python-docx==1.1.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.1.1
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.38.6
sympy==1.13.1
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.30.6
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
yarl==1.18.3
